requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
pydub==0.25.1
python-decouple==3.8
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.31.0
SpeechRecognition==3.10.0
tqdm==4.66.1
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
